Extract callCompletionApi.
